Finalized train and evaluation

+++ llava/model/language_model/llava_llama.py
@@ -88,7 +88,8 @@ class LlavaLlamaForCausalLM(LlamaForCausalLM, LlavaMetaForCausalLM):
                 image_sizes
             )
         # TODO: Double check the following modification you made. Without this it would fail due to type mismatch :(
-        inputs_embeds = inputs_embeds.to(torch.bfloat16)
+        # TODO: For evaluation this is not required! Super wierd
+        # inputs_embeds = inputs_embeds.to(torch.bfloat16)
         return super().forward(
             input_ids=input_ids,
             attention_mask=attention_mask,
